+++ src/mobact.c
@@ -1336,7 +1336,7 @@ void scale_mob_to_level(char_data *mob, int level) {
 	mob->points.max_pools[MANA] = MAX(0, (int) ceil(value));
 	
 	// blood*
-	value = size_data[(int)GET_SIZE(mob)].max_blood;
+	value = size_data[GET_SIZE(mob)].max_blood;
 	value += MOB_FLAGGED(mob, MOB_VAMPIRE) ? ((10 * high_level) + (20 * over_level)) : 0;
 	value *= MOB_FLAGGED(mob, MOB_ANIMAL) ? 0.5 : 1.0;
 	mob->points.max_pools[BLOOD] = MAX(1, (int) ceil(value));
